+++ lang/en/layouts.php
@@ -3,7 +3,7 @@
 return [
 
     'identity-link' => '<a href=":url">Connect an identity</a> to your account to sign in with a single click.',
-    'use-app' => 'For your convenience and security, you can use this service as an app.',
+    'use-app' => 'For your convenience and security, you can <a href="#" role="button" data-add-to-homescreen>use this service as an app</a>.',
     'app-login' => 'You can sign in for more options.',
 
     'login.description' => 'Log in with a provider or request a verification code.',